Move client version query from adaptor to client
